Added Playback and Paused modes
Added the ability to load the "best bird", save the current running state, and playback the bird with that brain until it dies.
Also added the ability to pause on P.
In addition, the "best bird" is saved to a cookie rather than a file, and that cookie is loaded (if it exists), at startup. Otherwise the first bird ever generated becomes the best one.